+++ gDRplots/NEWS.md
@@ -1,3 +1,6 @@
+## gDRplots 0.0.28 - 2024-08-22
+* update heatmap function to return plot and data - change file schema
+
 ## gDRplots 0.0.27 - 2024-08-20
 * move functions related to interactive plots to the `gDRiPlots` package